Finished preprocessing task
It should be working as desired now. As always please double check but
according to print(X.ndim) and print(y.ndim) they are a 2 and 1
dimensional arrays which is what we needed. X now counts from 1 to 569
and the other values in the array are the real valued ones